update README.md with compatible versions for tfq
Currently, due to a Google Colabs update, the libraries are not compatible with the current version of a few dependencies. This description in the README.md document includes compatible versions for the following packages: Cirq, Tensorflow, Python, and Tensorflow Quantum.

The configuration was tested and successfully ran on a MacBook Air 1.6 GHz Dual-Core Intel Core i5 (macOS 14.0).