Fix compiler warnings about unused variables
This file produces many warnings of the form

```
tensorflow_quantum/core/src/circuit_parser_qsim.cc:185:8:
warning: variable 'unused' set but not used [-Wunused-but-set-variable]
  185 |   bool unused = absl::SimpleAtoi(op.qubits(0).id(), &q0);
      |        ^
```

Inspecting the code reveals that variable `unused` is indeed set but
never used. I replaced the code with simple testing of the return
values, similar to what is used in one place in the file, and in a way
that follows the recommendations from the Abseil docs.