Fix "unused lambda capture" compiler warnings
One of the cases of lambda capture in this file resulted in warnings
of this sort:

```
tensorflow_quantum/core/ops/tfq_simulate_state_op.cc:154:29:
warning: lambda capture 'max_num_qubits' is not used [-Wunused-lambda-capture]
  154 |       auto copy_f = [i, nq, max_num_qubits, &output_tensor, &ss, &sv](
      |                           ~~^~~~~~~~~~~~~~
```

The parameter really was not used in the body. Removing it from the
lambda capture silenced the warning.