Silence compiler warnings we can't do anything about
A number of harmless warnings during compilation come from third-party
things pulled in by the TFQ build rules. They involve things like
unused functions, unused variables, or (in the case of zlib) some
deprecated uses of prototypes. The best way to fix them would be in
the original source code, but in some cases this is impossible; for
example, we can't change what TensorFlow 2.15.0 needs. The second best
way would be to add patch rules in TFQ to adjust the code or the build
rules. I spent a couple of hours trying to figure out how to that for
the copy of zlib included by bazel_rules, but due to my still-limited
Bazel knowledge, I wasn't able to get the patching scheme working. So,
for the time being, and in consideration of the harmless nature of the
warnings, the approach taken here is to add compile-time per-file
flags that tell the compiler not to warn about the specific kinds of
problems that we know exist in the files involved.