Update version of nbformat used
In more recent versions of Python like 3.10, the `base64` package
changed APIs, and this in turn caused nbformat 4.4.0 to fail to work.
Updating to the latest version (5.1.3) solved that problem.

A minor tweak was also needed in scripts/format_ipynb.py to account
for a different error class being returned when "!pip ..." commands
are encountered in the notebook file.