Revert a change from commit f64f1b3
One of the `size_t` → `int` changes was wrong and led to a test
failure.

I would prefer to edit the commit itself; however, that would require
force-pushing to the TFQ repo, and I'm not sure if that will break the
code review comments in the PR in GitHub. So I'm making the change in
a separate commit.