Update .pylintrc to work with pylint version 3.3.3
This is a `.pylintrc` file that works with pylint 3.3.3. It started life
as the config file produced by pylint 3.3.3 with the
`--generate-rcfile` option. I then compared it to the previous
`.pylintrc` and adjusted the settings to match, followed by a long
period of test-adjust-retest:

1. run `pylint` on the existing code base
2. look at the warnings, find the relevant flags & turn them off
3. if there are more warnings, go back to step 1

This `.pylintrc` file thus has a lot of pylint checkers commented out.
We should systematically re-enable the checkers and fix the warnings
in the code. However, that should be left to separate PRs. This PR
only makes the minimal changes needed to use pylint 3.3.3.