Fix #887: pass a list to random.sample()
Using a set in a call to the `sample(…)` method from the Python `random`
standard library has been deprecated since Python 3.9, and in Python
3.11, it causes an error. This was happening in the function
`util.py:_apply_random_control(…)` starting on line 111. Changing the
set to a list in the call is probably the most efficient way of
addressing this compatibility problem, and is backward-compatible with
Python versions prior to 3.11.

I searched but could not find another place where a set was being passed
to `sample()`. (There are other calls to `random.sample()`, but
none involve sets, as far as I have been able to determine.)