Make the question about CPU be more clear
I've always the phrasing of that question confusing and misleading. (The
question is _really_ about GPU vs CPU, but it only asks about CPU and
does not say anything about the implications of replying with `n`.) We
may as well take this opportunity to try to make it more clear.