Use newer Bazel `rules_python` and read `requirements.txt`
Running in an environment where the system Python interpreter is not the
desired one continued to be a problem in my tests. This updates
WORKSPACE to use a slightly newer version of the rules_python library
than what is installed by TensorFlow's various workspace files, and in
addition, to make it read the `requirements.txt` file.