Switch to using `requirements.in` and `pip-compile`
The Python environment that Bazel sets up for `bazel test` lacked all
the dependencies needed. In addition, it's generally considered a good
practice to pin the versions of dependencies in order to get more
reproducible builds. TensorFlow has been using `pip-compile` and
`requirements_lock_X.Y.txt` files for this, where `X.Y` is the Python
version. This commit changes the TensorFlow Quantum scheme to be
similar: a `requirements.in` lists the essential requirements, and this
file is then processed by `pip-compile` to produce the final
`requirements.txt` by following the transitive dependencies.