Delete unneeded software to save space on runner
Deleting some of the pre-installed software directories on the runner
can save a lot of space. Deleting the Dotnet and Android installations
in particular saves ~13 GB. This will hopefully be enough to get past
the current problem with builds exceeding space on the runner.

Note: the recursive `rm` is backgrounded because otherwise, that step
takes 50-60 s. Doing it in the background lets other things proceed.

There was a risk that the background process would slow down the rest of
the workflow more than the 60s it would take to stop and wait for the
`rm` to finish. Thankfully, there's no overall impact: it turns out that
the overall workflow takes just as long as before. So, we get the
benefit of more space without run-time cost.