Don't print destination path at the end
When running this inside a Docker container for building a wheel, the
directory this script prints refers to paths inside the container, not
outside in the developer's environment. This can be very confusing;
moreover, it's not really necessary because (1) the path is given on the
command line and (2) this script prints the destination when it starts.