Simplify run_auditwheel.sh
The script in third_party/tf/auditwheel does not seem necessary (and
TensorFlow 2.16 doesn’t even have such a script anymore). As far as the
munging of the policy file that third_party/tf/auditwheel was doing,
adding `--exclude libtensorflow_framework.so.2` to auditwheel appears to
be enough to solve the problem. (At least, the resulting wheel has
worked in testing on Colab.)