Use `linux-x86-n2-32` runners to speed up some jobs
Thanks to the help of Google's ML Velocity team, the TensorFlow Quantum
project has access to larger job runners. We can use them for the really
time-consuming jobs in our workflows.

It took a fair amount of trial-and-error testing to resolve some odd
differences in the runner environments, but eventually I got it down to
just a couple of additional commands.

Note: I didn't change the `wheel-build` job to use the new runners
because the most time-consuming job in here is the tutorials tests and
the wheel-build job is not the bottleneck. The larger runners are more
expensive, so if we can't benefit from them, it doesn't make sense to
use them.