Remove `set -x` & a print stmt that leads to confusion
The final print statement about the directory is misleading when
`build_pip_package.sh` is being run inside a Docker container. On
balance, I think it's better to remove it. Hopefully the user running
the script will remember that they supplied the destination parameter as
an argument on the command line to begin with.