Delay deleting existing `.bazelrc` file
Delay removing the `.bazelrc` files until configure.sh is actually ready
to start writing a new one. That way, if the user ^C's in the middle of
configure.sh running, it's less likely to overwrite what they currently
have. (It happened to me: I realized I wanted to stop and check
something at the point when it asked for cpu vs gpu, but by then it was
too late: the .bazelrc file was gone.)