Add a step to run `git config blame.ignoreRevsFile`
In preparation for a separate PR to do nothing but reformat a few files,
this adds information telling developers how to configure git to ignore
specific commits. (This is the same approach used in Cirq.)

This also adds a placeholder `.git-blame-ignore-revs` file for now, so
that running the git config command does not result in an error.