Add support for a requirements constraint file
Some of the TFQ transitive dependencies have recently gotten Dependabot
alerts to be updated for some security flaws. However running
pip-compile again does not end up updating those dependencies. It seems
that the recommended way to handle this is to use a constraint file.

This updates `./scripts/generate_requirements.sh` to invoke pip-compile
with a constraint file.

It also tells pip-compile to write the name of this script into the
header of the requirements.txt file it generates.