Remove python_version constraints & nonessential requirements
The python_version actually means the constraint only applies to that
Python version. We need the packages installed no matter the Python
version; we just need the constraint on Python < 3.11. It's simpler and
still works to not add the python_version, so let's just remove them.

In addition, this removes constraints that should belong elsewhere, to
keep the requirements.txt file leaner.