Add pins file handling to generate_requirements & simplify
A shortcoming of generating requirements.txt with version pins is that
the versions of some transitive dependencies may need to be updated to
address security releases. Simply re-running pip-compile to produce a
new requirements.txt may not update the versions because the
dependencies that bring in the transitive dependencies have not
themselves changed. The recommended practice for this situation is to
put the version constraints into a separate constraint file and pass it
as an argument to pip-compile. (pip-compile has explicit support for
this.) Some additions to the script scripts/generate_requirements.sh
encapsulate and document this process.

In addition, I removed some no-longer-needed post-processing of the
requirements.txt file. It was a hack in the first place, and with the
latest requirements.in, it turns out to be unnecessary.

Finally, I simplified this some more and made it follow more style
guidelines for Bash scripts.