Unpin version of gym in the quantum reinforcement learning tutorial (#943)
Loading a specific version of gym in Colab causes Colab to require
another restart. In Colab, the newest version of gym is preinstalled.
That version of gym seems to work for the tutorial.

Note: the `import gym` step generates a warning that gym is no longer
supported. This is potentially confusing to users. The warning also
suggests switching to a different package (Gymnasium). I opened issue
#944 to remind us of this.