Add import of `seaborn` to mnist tutorial
`docs/tutorial/mnist.ipynb` imports the `seaborn` package, but does not
install it. It seems to come preinstalled in Colab and the DevSite
toolchain, but someone running the notebook outside of those
environments may hit a "module not found error". By adding a `!pip
install seaborn` after installing TF and TFQ, we can not only save users
the annoyance; we can also save future TFQ maintainers the time to debug
the problem if they encounter the problem during local testing.